Add files via upload
11. ADTs II & Data Structures
11.1 27-Way Trees
The Dictionary Abstract Data Type (ADT) allows words to be stored in such that they may be
efficiently checked later, to ensure that words have been spelled correctly etc. There are many
ways of implementing the Dictionary ADT (binary trees, hash tables and so on), but here we
look at a tree structure which has a dynamic collection of nodes.
The 27-Way Tree has 27 downward links, each corresponding to one of the letters a. . . z and one
to the apostrophe character '. This has many similarities with tries:

Exercise 11.1.1 Write an ADT to implement the t27.h interface given. Write the source
file t27.c such that the driver file driver.c works correctly. Ensure that this all works with
the Makefile which, as with t27.h must be used unaltered. The standard operations are :
dict* dict_init(void);
bool dict_addword(dict* p, const char* str);
dict* dict_spell(const dict* p, const char* str);
int dict_nodecount(const dict* p);
int dict_wordcount(const dict* p);
int dict_mostcommon(const dict* p);
void dict_free(dict** p);
The function dict_addword() allows a string to be inserted into the tree, and the node after
marked as a terminal node.
dict_spell() returns a pointer to the terminal node corresponding to the string (if it has
been inserted) and NULL otherwise.
dict_nodecount() returns the total number of nodes which are part of the tree, and
dict_wordcount() returns the total number of words which have been inserted into the tree
including duplicates. A frequency counter is used in each node to keep a track of words
which have been added multiple times. In this case this counter is incremented, but no new
nodes are created.
dict_mostcommon() returns the number of times the most common word has been inserted.
The above functions are the standard assignment and are worth 60%. There are two additional
advanced functions for you to implement, each worth up to 10% of the assignemnt:
unsigned dict_cmp(dict* p1, dict* p2);
which counts the least number of nodes you must traverse to move from one node to another
in the tree. In the above figure, there are 7 steps from ‘car’ to ‘part’. The other is:
void dict_autocomplete(const dict* p, const char* wd, char* ret);
which returns the substring corresponding to the additional letters required to get from the wd
to the most frequent word below it. In the figure, to autocomplete ‘car’ would be the word
‘cart’, so the additional letters required are simply ‘t’, which is stored in ret.
• We’ve provided the structure that must be used for this assignment. Do not attempt to
change it. All functionality required is possible using this structure.
• For questions such as "but what about hyphenated words?" etc., look in the driver.c
file. If we haven’t tested for it in there, make a sensible decision as to what should
happen. Provided that this file still operates correctly you can decide some of the
edge-cases for yourself.
• Even if you don’t get all the functions to work correctly, make sure the code still
compiles by writing ‘dummy’ functions as placeholders, even if some of the assertions
fail. I’ll test the basic functionality separately from the two advanced ones.
• This assignment is brand new. If minor typos etc. are uncovered as you complete the
assignment we may update this documentation and/or the online files provided. Please
check regularly to make sure you are using the most recent version.